Teach pytest_test and pylint_test `shard_count`
See https://docs.bazel.build/versions/master/test-encyclopedia.html#test-sharding
for the explanation, and this github query
https://github.com/search?q=TEST_SHARD_INDEX&type=code which shows
many examples of how to use it.

For pytest we simply register a plugin which uses
https://docs.pytest.org/en/stable/reference.html#pytest.hookspec.pytest_collection_modifyitems
to filter the selected tests that run per shard.

For pylint we have to wrap the main entrypoint and filter the _source_
_files_ that each shard gets. Pylint has no more granular filtering,
which is okay. We have some directories that contain a _lot_ of source
files and putting them all in one `pylint_test` makes it slow. It can
also make iterative development slow because pylint doesn't seem to have
any ability to do per-file caching
https://mail.python.org/pipermail/code-quality/2015-July/000621.html