Fix --test_arg with pylint_test
Before, pylint_main expected to find only sources after a `"--"`
separator on the command line. If a user specified `--test_arg` to
bazel then it would be treated like a source file and the flag would
only apply to one shard.

Now the source files are templated into the pylint_main file so there is
no mucking around with cli flags.