Switch python_interpreter based on the platform
For the same reason we switch requirements.txt we need to switch the
actual interpreter.

We setup our toolchain so that even if bazel is being run from within a
virtualenv it'll still use the toolchain python. This breaks when we
pass `python_interpreter = PYTHON3`. Before we only did this when
compiling dependencies, but now it's also done when fetching them so
it's more visible.