py_unzip: Create an archive that can be expanded into a standalone, runnable python application.
The Problem:

We store two copies of evey python file and c-extension in `parfile`s/the `unpar` directory.

---

The Current State of Affairs:

Take the following example `BUILD` file,
```py
// BUILD.bazel
pkg_python_app(
  name = "my_app",
  srcs = ["my_app.py"],
  bindir = "/usr/bin",
  libdir = "/lib",
  tar = "staging",
)
```
This created a `staging.tar` which when extracted created the following files,

```
/usr/bin/my_app  # adds environment variables and calls '/lib/par/my_app.par'
/lib
  /par/my_app.par  # a parfile (zip,) containing all dependencies
  /unpar/my_app/ # the extracted version of 'my_app.par'
    __main__.py
    PAR_MANFIEST
    pypi__36__attrs_20_3_0/...
    com_128technology_i95/...
```

`/lib/unpar/my_app` is created the first time the application is run, or could be created eagerly by runing the executable with the environment variable `PAR_EXTRACT_ONLY=1`.

The `parfile` had no compression applied to it, so that means **we keep two copies of every single python file and c-extension**.

---

The History:

zipfiles created by bazel (via the commandline flag `--build_python_zip`) cannot be run outside of the build/ directory.
We chose `par_binary` (`google/subpar`) because it creates `parfile`s which _can_ be copied anywhere and executed (the python interpreter is not bundled in the parfile.)

parfiles work best with `zip_safe = True`.
That allows the python interpreter to find the python source files directly within the parfile without having to extract the archive first.
c-extensions are **not** `zip_safe` so any `parfile` that includes one must be marked as `zip_safe = False`.
Most of our python applications include a c-extension as a dependency (`lxml`, `pycrypto`, etc...) so _most_ are marked as `zip_safe = False`.

`google/subpar` extracts to a **unique** tempdir when `zip_safe = False`.
This lead to extremely large overhead for python "scripts" that were supposed to be invoked, do one thing, and exit.
For example, we were running thousands of `salt` commands and each one was taking 5+ seconds to extract _before_ the beginning of `main()` was even called.

We forked and created `128technology/subpar` which added a new `extract_dir` argument.
This allowed the `parfile` to extract to a Known Good Place and re-use the extracted version of the application as long as the parfile didn't change.
The `parfile` used a hash of all the files in the archive to create the `PAR_MANFIEST` so as to not re-extract unless that file in the extracted dir didn't match the one embeded within the `parfile`.
The `parfile` still contained the `__main__.py` so the `exec_wrapper` called the `parfile` which used all the files in `unpar/` _except_ the `__main__.py`.

We then continued to add more patches on top of `subpar` to byte-compile eagerly, etc...

---

The Solution:

`subpar` has outlived it's usefulness. We've hacked it to bits so it's time to re-write the parts we use from the ground(ish) up.

Meet `py_unzip`.

Take the same `BUILD` file from before and add `use_py_unzip = True`
```py
// BUILD.bazel
pkg_python_app(
  name = "my_app",
  srcs = ["my_app.py"],
  bindir = "/usr/bin",
  libdir = "/lib",
  use_py_unzip = True,  # <--- new!
)
```

Now `my_app.tar` is created which extracts to,
```
/usr/bin/my_app  # adds environment variables and calls '/lib/unzip/my_app/__main__.py'
/lib/unzip/my_app/
  __main__.py
  runfiles/
    pypi__36__attrs_20_3_0/...
    com_128technology_i95/...
```
There is no duplication between the archive/unarchived version of the application.

---

The Implementation:

- Take the zipfile that bazel generates via `py_binary` (same as `--build_python_zip`.)
  This is done by `ctx.attr.src[OutputGroupInfo].python_zip_file` which is the same as
```
filegroup(
  name = "foo_zip",
  srcs=  ["//bar:foo"],
  output_group = "python_zip_file",
)
```
  as mentioned in this github comment, https://github.com/bazelbuild/bazel/issues/3530#issuecomment-536454925.

- Replace the `__main__.py` with one that works in the extracted directory.
  Bazel's main comes from `tools/python/python_bootstrap_template.txt` which templates in a half-dozen variables; our new `__main__.py` is based on that, but simplified to only work for our use-case (extracted outside of bazel.)