Don't pass PYTHONPATH down to subprocesses
---
The Problem?

Starting with a quick recap of what a `py_unzip` actually is.
Given the following BUILD file,

```bzl
py_unzip(
  name = "my_app",
  srcs = ["__main__.py"],
  main = "__main__.py",
  libdir = "/unzip",
)
```

Running `bazel build //path/to/my_app:my_app.tar` generates a tarfile that contains the following files,

```
  /unzip/
    __main__.py  # This is __main__.py.tmpl with substitutions. Always named `__main__.py`.
    runfiles/<workspace-name>/path/to/my_app/
      __main__.py  # This is the `main` argument to the `py_unzip` macro. Can be named whatever the user specified.
```

`py_unzip` works by generating a wrapper `__main__.py` (using `__main__.py.tmpl`) that wraps the "real" `py_unzip.main` attribute.
The primary purpose of the wrapper is to setup the import path (via the `"PYTHONPATH"` environment variable,) and `execv` the real main.

This poses issues when the python application subprocesses another python process that has it's own python path.
`subprocess.run` inherits all the parents' environment variables so the subprocess will inherit the `"PYTHONPATH"` of the `py_unzip` application.
`"PYTHONPATH"` also has a higher precedence then site-packages so the `py_unzip`s dependencies will be found first.
This leads to errors either via name conflicts or different versions of the same package.

---
The Solution?

Inject code into the user-specified `main` file that unsets `"PYTHONPATH"`.
This code is inserted at the beginning of the main file so it's the first code that executes, even before other `import` statements.

This works because when the python interpreter starts it reads the `"PYTHONPATH"` only once and saves that into `sys.path` which is what's used for looking up packages when importing.
That means that changing the `"PYTHONPATH"` doesn't impact the current interpreter's ability to `import`.

---
Technical Details

The changes to `rezipper.py` models `google/subpar`s compiler which inserts a similar "runtime setup" at the beginning of the `__main__.py`.
The difference is that `subpar`'s runtime is much more complex (like unpacking the zipfile dynamically,) so they ship an entire package.